requirements.txt
opencv-python>=4.8.0
pyserial>=3.5 